整合了HashTable的实现

+++ src/util/KVStore.cpp
@@ -1,8 +1,9 @@
 #include "KVStore.h"
-
+#include "../resp/RespValue.h"
+#include "HashTable.h"
 template <typename T>
 KVStore<T>::KVStore(size_t shardCount) : shardCount(shardCount) {
-    for (size_t i = 0; i < shardCount; i++) {
+    for (int i = 0; i < shardCount; i++) {
         shards.emplace_back(std::make_unique<Shard>());
     }
 }
